Refactor web_analytics and create utils.ts
