+++ portal/server/package.json
@@ -8,6 +8,7 @@
         "start": "next start"
     },
     "dependencies": {
+        "@amplitude/analytics-node": "^1.3.6",
         "@sentry/nextjs": "^8",
         "@sentry/node": "^8.49.0",
         "@vercel/analytics": "^1.4.1",
